consolidate exports through index files
