+++ src/background/settings/index.js
@@ -1,0 +1,3 @@
+export { quickStorage } from './quickStorage.js';
+export { saveSettings } from './save.js';
+export { default as defaultSettings } from './defaultSettings.json';
